refactor: rename package entrypoint file
