+++ psql2bigquery/entrypoint.py
